Faster implementation
Rather than wait until the tmpSrc is loaded before loading the src
we are now instantiating a virtual Image object and setting its src to
the target src path. This will async load the src. All events are
attached to the virtual object and handled there.

tldr; cleaner and faster